+++ lib/authrocket/realm.rb
@@ -9,8 +9,10 @@ module AuthRocket
     has_many :orgs
     has_many :users
 
-    attr :api_key_policy, :api_key_prefix, :custom, :name
-    attr :require_unique_emails, :state, :username_validation_human
+    attr :api_key_policy, :api_key_prefix, :custom, :name # :api_key_minutes
+    attr :jwt_data, :require_unique_emails, :session_minutes, :session_type
+    attr :state, :username_validation_human
+    attr :jwt_secret # readonly
 
 
     def reset!(params={})